chore: Add ExtractorService to handle query extraction
