+++ templates/types/extractor/fastapi/app/models/request.py
@@ -1,0 +1,13 @@
+from pydantic import BaseModel
+
+
+class RequestData(BaseModel):
+    query: str
+
+    class Config:
+        json_schema_extra = {
+            "examples": [
+                {"query": "What's the maximum weight for a parcel?"},
+            ],
+        }
+
